techtest_mleng-master-ans/requirements.txt
flask==1.1.2
numpy==1.19.4
requests==2.25.0
gunicorn==20.0.4
python-dotenv==0.15.0